MDEV-13073. This part refines the semisync ack receiver thread start and termination.
The ack receiver thread termination specifically is turned
to follow a standard pattern. The ack thread (child) and the stopping
thread (parent) communicate through a state variable and signal/wait-for
calls. Such pattern also covers windows execution.
Necessary changes are done to signatures of THD::THD,init to
propage a fact of already locked LOCK_global_system_variables (it
can't be unlocked as such alternative is prone to deadlock).